+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Internal
 
+* Factor out internal interface ManagedObject
 
 ## 3.3.1 (2017-05-26)